prv_api: group serializers in a single file
