+++ mds/apis/prv_api/providers.py
@@ -1,32 +1,10 @@
-from rest_framework import serializers
 from rest_framework import viewsets
 
 from mds import models
 from mds.access_control.permissions import require_scopes
 from mds.access_control.scopes import SCOPE_PRV_API
 
-
-class ProviderSerializer(serializers.ModelSerializer):
-    id = serializers.UUIDField(
-        required=True, help_text="Unique provider identifier (UUID)"
-    )
-    name = serializers.CharField(help_text="Name of the Provider")
-    logo_b64 = serializers.CharField(
-        required=False, help_text="Logo of provider base64 encoded"
-    )
-
-    class Meta:
-        model = models.Provider
-        fields = (
-            "id",
-            "name",
-            "logo_b64",
-            "base_api_url",
-            "device_category",
-            "api_authentication",
-            "api_configuration",
-            "agency_api_authentication",
-        )
+from .serializers import ProviderSerializer
 
 
 class ProviderViewSet(viewsets.ModelViewSet):
